requirements-mac.txt
# macOS-specific dependencies
# Установка: pip install -r requirements-mac.txt
# Предварительно: brew install blackhole-2ch

# Core
python-dotenv>=1.0.0

# Захват системного аудио (через BlackHole виртуальное устройство)
sounddevice>=0.4.6

# Транскрипция (Apple Silicon / MLX)
mlx-whisper>=0.3.0

# Диаризация
pyannote.audio==4.0.4
# PyTorch: pyannote 4.x тянет 2.11+, torchaudio опциональна (pyannote 4.x не использует AudioMetaData)

# LLM (опционально — через Claude CLI, не через SDK)
# anthropic>=0.25.0

# Tray (menu bar)
pystray>=0.19.5
Pillow>=10.0.0

# Уведомления
plyer>=2.1.0
